[refactor]: refactor rds tests
